# backend/internal/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "weflow swagger api documention.",
        "title": "weflow",
        "contact": {
            "name": "hertz-contrib",
            "url": "https://github.com/hertz-contrib"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/weflow",
    "paths": {
...
        "/model/detail/get": {
            "get": {
                "description": "获取模版详情(模板信息，版本，流程定义，流程定义详情)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模板"
                ],
                "summary": "获取模版详情(模板信息，版本，流程定义，流程定义详情)",
                "parameters": [
                    {
                        "type": "string",
                        "example": "",
                        "description": "模型id",
                        "name": "modelID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "",
                        "description": "版本id",
                        "name": "versionID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bo.ModelDetailResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/model/details": {
            "post": {
                "description": "获取所有组的所有模版",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模板"
                ],
                "summary": "获取所有组的所有模版",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "GroupModelQueryVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.GroupModelQueryVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bo.GroupModelDetailsResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/model/group/add": {
            "post": {
                "description": "添加模板组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模板"
                ],
                "summary": "添加模板组",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.ModelGroupAddVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/model/group/del": {
            "post": {
                "description": "删除模板组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模板"
                ],
                "summary": "删除模板组",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "ModelGroupDelVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.ModelGroupDelVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/model/group/edit": {
            "post": {
                "description": "编辑模板组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模板"
                ],
                "summary": "编辑模板组",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "ModelGroupEditVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.ModelGroupEditVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
...
        "/model/invalid": {
            "post": {
                "description": "停用模板",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模板"
                ],
                "summary": "停用模板",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "ModelInvalidVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.ModelInvalidVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/model/list": {
            "post": {
                "description": "获取模板列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模板"
                ],
                "summary": "获取模板列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "ModelQueryVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.ModelQueryVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bo.ModelDetailResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/model/page": {
            "post": {
                "description": "分页获取模板列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模板"
                ],
                "summary": "分页获取模板列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "ModelPageVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.ModelPageVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bo.ModelDetailResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/model/publish": {
            "post": {
                "description": "发布模板",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模板"
                ],
                "summary": "发布模板",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "ModelSaveVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.ModelSaveVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/model/save": {
            "post": {
                "description": "保存模板",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模板"
                ],
                "summary": "保存模板",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "ModelSaveVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.ModelSaveVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/model/version/get": {
            "get": {
                "description": "根据模板查询获取模板版本列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模板"
                ],
                "summary": "根据模板查询获取模板版本列表",
                "parameters": [
                    {
                        "type": "string",
                        "example": "",
                        "description": "模型id",
                        "name": "modelID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bo.ModelVersionResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/model/version/release": {
            "post": {
                "description": "上线模板版本(模板版本列表启用版本)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模板"
                ],
                "summary": "上线模板版本",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "ReleaseModelVersionVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.ReleaseModelVersionVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "base.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "0:成功，其他：失败",
                    "type": "integer"
                },
                "data": {
                    "description": "数据"
                },
                "msg": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
        "bo.GroupModelDetailsResult": {
            "type": "object",
            "properties": {
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "createUser": {
                    "description": "创建人",
                    "type": "string"
                },
                "groupID": {
                    "description": "组id",
                    "type": "string"
                },
                "groupName": {
                    "description": "组名称",
                    "type": "string"
                },
                "id": {
                    "description": "唯一id",
                    "type": "integer"
                },
                "models": {
                    "description": "模型列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bo.ModelDetailResult"
                    }
                },
                "remark": {
                    "description": "描述",
                    "type": "string"
                },
                "updateTime": {
                    "description": "更新时间",
                    "type": "string"
                },
                "updateUser": {
                    "description": "更新人",
                    "type": "string"
                }
            }
        },
...
        "bo.ModelDetailResult": {
            "type": "object",
            "properties": {
                "createTime": {
                    "description": "创建时间 yyyy-MM-dd HH:mm:ss:SSS",
                    "type": "string"
                },
                "createUser": {
                    "description": "创建人",
                    "type": "string"
                },
                "formDefID": {
                    "description": "表单定义id",
                    "type": "string"
                },
                "iconURL": {
                    "description": "icon图标地址",
                    "type": "string"
                },
                "id": {
                    "description": "唯一id",
                    "type": "integer"
                },
                "modelGroupID": {
                    "description": "模版组id",
                    "type": "string"
                },
                "modelID": {
                    "description": "模板id",
                    "type": "string"
                },
                "modelTitle": {
                    "description": "模板标题",
                    "type": "string"
                },
                "processDefID": {
                    "description": "流程定义id",
                    "type": "string"
                },
                "remark": {
                    "description": "描述",
                    "type": "string"
                },
                "status": {
                    "description": "模板状态【1：草稿；2：发布；3：停用】默认草稿",
                    "type": "integer"
                },
                "updateTime": {
                    "description": "更新时间 yyyy-MM-dd HH:mm:ss:SSS",
                    "type": "string"
                },
                "updateUser": {
                    "description": "更新人",
                    "type": "string"
                }
            }
        },
...
        "vo.GroupModelQueryVO": {
            "type": "object",
            "properties": {
                "modelName": {
                    "description": "模型名称",
                    "type": "string",
                    "example": ""
                }
            }
        },
...
        "vo.ModelAdvancedSetup": {
            "type": "object",
            "properties": {
                "titleContent": {
                    "description": "标题内容",
                    "type": "string",
                    "example": ""
                },
                "titleType": {
                    "description": "标题类型【1：默认；2：自定义】默认为1",
                    "type": "integer",
                    "example": 0
                }
            }
        },
...
        "vo.ModelGroupAddVO": {
            "type": "object",
            "properties": {
                "groupName": {
                    "description": "组名称",
                    "type": "string",
                    "example": ""
                },
                "remark": {
                    "description": "描述",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "vo.ModelGroupDelVO": {
            "type": "object",
            "properties": {
                "groupID": {
                    "description": "组id",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "vo.ModelGroupEditVO": {
            "type": "object",
            "properties": {
                "groupID": {
                    "description": "组id",
                    "type": "string",
                    "example": ""
                },
                "groupName": {
                    "description": "组名称",
                    "type": "string",
                    "example": ""
                },
                "remark": {
                    "description": "描述",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "vo.ModelInvalidVO": {
            "type": "object",
            "properties": {
                "modelID": {
                    "description": "模型id",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "vo.ModelPageVO": {
            "type": "object",
            "properties": {
                "modelName": {
                    "description": "模型名称",
                    "type": "string",
                    "example": ""
                },
                "pageNum": {
                    "description": "页码",
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "description": "每页条数",
                    "type": "integer",
                    "example": 30
                },
                "status": {
                    "description": "模板状态【1：草稿；2：发布；3：停用】默认草稿",
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "vo.ModelQueryVO": {
            "type": "object",
            "properties": {
                "modelName": {
                    "description": "模型名称",
                    "type": "string",
                    "example": ""
                },
                "status": {
                    "description": "模板状态【1：草稿；2：发布；3：停用】默认草稿",
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "vo.ModelSaveVO": {
            "type": "object",
            "properties": {
                "advanced": {
                    "description": "高级设置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.ModelAdvancedSetup"
                        }
                    ]
                },
                "base": {
                    "description": "基础设置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.ModelBaseSetup"
                        }
                    ]
                },
                "flowContent": {
                    "description": "流程内容",
                    "type": "string",
                    "example": ""
                },
                "formContent": {
                    "description": "表单内容",
                    "type": "string",
                    "example": ""
                },
                "modelID": {
                    "description": "模型id",
                    "type": "string",
                    "example": ""
                }
            }
        },
...
        "vo.ReleaseModelVersionVO": {
            "type": "object",
            "properties": {
                "versionID": {
                    "description": "版本id",
                    "type": "string",
                    "example": ""
                }
            }
        },
...
    }
}